Refactored Role Summary and added to bottom of page
