--- src/components/presentational/ApplyNow/styles.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-import { JobNavSubcomponentLayout } from '../../container/JobDetails/styles.js'
-
-export const ApplyNowTag = styled.div`
-  width: 100%;
-  background-color: white;
-`
